+++ spec/integrations/instrumentation/error_callback_multiple_subscription_groups.rb
@@ -6,7 +6,7 @@
 
 setup_karafka do |config|
   # Bad port on purpose to trigger the error
-  config.kafka = { 'bootstrap.servers' => '127.0.0.1:9090' }
+  config.kafka = { 'bootstrap.servers': '127.0.0.1:9090' }
 end
 
 draw_routes do