+++ spec/spec_helper.rb
@@ -58,7 +58,7 @@ module Karafka
   # Configuration for test env
   class App
     setup do |config|
-      config.kafka = { 'bootstrap.servers' => '127.0.0.1:9092' }
+      config.kafka = { 'bootstrap.servers': '127.0.0.1:9092' }
       config.client_id = rand.to_s
       config.pause_timeout = 1
       config.pause_max_timeout = 1