+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 - ActiveJob Rails adapter
 - Added benchmarks that can be used to profile Karafka
 - Standardize error hook for all error reporting
+- Require `kafka` keys to be symbols
 
 ## 1.4.11 (2021-12-04)
 - Source code metadata url added to the gemspec